Fixed Unit test and compiling error after syncing with the upstream master; Grouped some router path for sharing; Moved getAuthHeader to httputil package
