+++ lib/backend/artifactorybackend/security/security.go
@@ -14,17 +14,12 @@
 package security
 
 import (
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"net/http"
-	"sync"
-	"time"
 
 	"github.com/docker/docker-credential-helpers/client"
 	"github.com/docker/engine-api/types"
 	"github.com/uber/kraken/utils/httputil"
-	"github.com/uber/kraken/utils/log"
 )
 
 const tokenUsername = "<token>"
@@ -36,24 +31,6 @@ type Config struct {
 	TLS                    httputil.TLSConfig `yaml:"tls"`
 	BasicAuth              *types.AuthConfig  `yaml:"basic"`
 	RemoteCredentialsStore string             `yaml:"credsStore"`
-}
-
-// RegistryToken is a bearer token to be sent to a registry
-type registryToken struct {
-	mux sync.Mutex
-	tokenHeader map[string]string
-	startTime time.Time
-	// validInterval in ns
-	validInterval int64
-}
-
-var regToken = registryToken{tokenHeader: make(map[string]string)}
-
-const tokenQuery = "http://%s/v2/token"
-
-type tokenAPIResponse struct {
-	Token string `json:"token"`
-	Expire uint32 `json:"expires_in"`
 }
 
 // GetHTTPOption returns httputil.Option based on the security configuration.
@@ -82,44 +59,6 @@ func (c Config) GetHTTPOption(addr, repo string) (httputil.SendOption, error) {
 		return nil, fmt.Errorf("basic auth: %s", err)
 	}
 	return httputil.SendTLSTransport(rt), nil
-}
-
-func GetAuthHeader(address string, opt httputil.SendOption) (map[string]string, error) {
-	URL := fmt.Sprintf(tokenQuery, address)
-
-	regToken.mux.Lock()
-	defer regToken.mux.Unlock()
-	currentTime := time.Now()
-	if len(regToken.tokenHeader) == 0 || (currentTime.Sub(regToken.startTime).Nanoseconds() >= regToken.validInterval) {
-		// Need to obtain a new token
-		resp, err := httputil.Get(
-			URL,
-			opt)
-		if err != nil {
-			return nil, fmt.Errorf("get token: %s", err)
-		}
-		defer resp.Body.Close()
-		body, err := ioutil.ReadAll(resp.Body)
-		if err != nil {
-			return nil, fmt.Errorf("read token response: %s", err)
-		}
-
-		token := tokenAPIResponse{}
-		err = json.Unmarshal(body, &token)
-		if err != nil {
-			return nil, fmt.Errorf("unmarshal token API response: %s", err)
-		}
-		regToken.tokenHeader["Authorization"] = token.Token
-		regToken.startTime = currentTime
-		regToken.validInterval = int64(token.Expire) * 1e+9
-		log.Infof("Update regToken %v", regToken)
-	}
-	// Deep copy to avoid race condition
-	res := make(map[string]string)
-	for k, v := range regToken.tokenHeader {
-		res[k] = v
-	}
-	return res, nil
 }
 
 func (c Config) getCredentials(helper, addr string) (types.AuthConfig, error) {
